wayfare-api-develop/wayfare-api-develop/requirements.txt
aniso8601==4.0.1
astroid==2.0.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
marshmallow==2.16.3
mccabe==0.6.1
pylint==2.1.1
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
rope==0.11.0
six==1.11.0
SQLAlchemy==1.2.14
urllib3==1.24.1
webargs==4.1.2
Werkzeug==0.14.1
wrapt==1.10.11
